Added code for the Tetrimino class which implied some modifications in others class + changing the type of the x y fields from float to int

+++ src/Coordinate.java
@@ -1,15 +1,15 @@
 
 public class Coordinate {
-	public float x;
-	public float y;
+	public int x;
+	public int y;
 	
-	public Coordinate(float x, float y) {
+	public Coordinate(int x, int y) {
 		this.x = x;
 		this.y = y;
 	}
 	
 	public Coordinate() {
-		this(0f, 0f);
+		this(0, 0);
 	}
 	
 	public Coordinate(Coordinate c) {
